refactoring1
