+++ step1/index.html
@@ -76,7 +76,7 @@
     </style>
   </head>
   <body>
-    <main id="app"></main>
-    <script type="module" src="./src/main.js" defer></script>
+    <app id="app"></app>
+    <script type="module" src="./src/app.js" defer></script>
   </body>
 </html>
