+++ step1/src/app.js
@@ -1,0 +1,3 @@
+import { render } from './service.js';
+
+render();
