+++ step2/README.md
@@ -8,16 +8,16 @@
     - [ ] ES5 vs ES6 차이점 정리
   - [ ] 파일을 기능 단위로 분리해본다.
     - [ ] core: 어플리케이션의 베이스 코드
-    - [ ] components: 컴포넌트 코드
+    - [x] components: 컴포넌트 코드
     - [ ] utils: 유틸리티 성향의 코드
-    - [ ] constants: 상수
-    - [ ] app.js (entry point)
-  - [ ] 다음과 같은 규칙을 지켜가며 코딩한다.
-    - [ ] 한 메소드(함수)에 indent(tab)는 최대 2depth로 유지하기
-    - [ ] else 예약어(keyword)를 쓰지 않는다.
-    - [ ] 상수를 적극적으로 사용한다.
-    - [ ] 한 줄에 점을 하나만 찍는다.
-    - [ ] 줄여쓰지 않는다 (축약 금지)
+    - [x] constants: 상수
+    - [x] app.js (entry point)
+  - [x] 다음과 같은 규칙을 지켜가며 코딩한다.
+    - [x] 한 메소드(함수)에 indent(tab)는 최대 2depth로 유지하기
+    - [x] else 예약어(keyword)를 쓰지 않는다.
+    - [x] 상수를 적극적으로 사용한다.
+    - [x] 한 줄에 점을 하나만 찍는다.
+    - [x] 줄여쓰지 않는다 (축약 금지)
 - [ ] 번들러 조사 및 적용
   - [ ] 번들러에 대해 알아보기
     - [ ] javascript 번들링
